portability fixes (mac os)
cado compiles on mac os x, core 2 !
read and copy Makefile.local.macos

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado/trunk@2008 de4796e2-0b1e-0410-91b4-86bc23549317

+++ linalg/bw/auxfuncs.c
@@ -97,7 +97,7 @@ static void enforce_limit(int l, rlim_t min, const char * name)
 
 	if (min==RLIM_INFINITY || lim.rlim_cur < min) {
 		die("FATAL: Limit %s is %d while we need at least %d\n",-31,
-				name,lim.rlim_cur,min);
+				name,lim.rlim_cur,(unsigned long) min);
 	}
 }
 
@@ -140,7 +140,7 @@ void coredump_limit(int allow)
 			die("getrlimit() failed!\n",31);
 		if (lim.rlim_max != RLIM_INFINITY)
 			fprintf(stdout,"// (actual limit is %lu bytes)\n",
-					lim.rlim_max);
+					(unsigned long) lim.rlim_max);
 		lim.rlim_cur=lim.rlim_max;
 		if (setrlimit(RLIMIT_CORE,&lim) < 0)
 			die("setrlimit() failed!\n",31);
